requirement.txt
Flask==3.1.0
Werkzeug==3.1.3
Pandas==2.2.3
Matplotlib==3.10.0
Seaborn==0.13.2
Statsmodel==0.14.4
scikit-learn==1.6.1
google.generativeai==0.8.4
dotenv==1.0.1
numpy==1.26.4